feat: 객체 데이터의 타입과 데이터 파일을 분리합니다.
